requirements.txt
-e git+https://github.com/openai/baselines.git@0961f5dd94fbf2a97688c17870fa3714773d180e#egg=baselines
torch==1.3.1
tqdm==4.35.0
numpy==1.17.2
Pillow==6.1.0
matplotlib==3.1.1
